ACP-L1-V2-E03-chat-sequencer: address PR #1769 blocking review
Fixes all 5 blocking findings from the 2026-08-04 review:
- AdvanceStreamHead now validates the requested position was actually
  committed by the stated source identity for this exact session
  (sessionRecord.sequencedPositions), rejecting fabricated/cross-session
  positions with a new UncommittedStreamPositionError; AcknowledgeAttachment
  no longer trusts a bare ReadOutcomeProgress as proof of a fully retained
  range.
- AdvanceStreamHead now checks ctx.Err() before mutating, matching
  Sequence/AcknowledgeAttachment.
- Sequence preserves the source-native payload byte-for-byte via a hand-spliced
  envelope instead of json.Marshal(item), which was silently compacting and
  HTML-escaping the nested json.RawMessage.
- Removed the WithEventsAppender/WithEventsReader secondary injectors;
  NewStore now takes both as direct constructor parameters.
- Consolidated stream_head.go/attachment_ack.go into sequencing.go and merged
  the pkg/wire identity test file into chat_sessions_composition_test.go,
  eliminating the lane-owned pkg-file-count growth (pkg/wire back to zero net
  growth; chat_sessions/internal/service reduced from 29 to 27 files).
