+++ pkg/services/chat_sessions/internal/service/logging.go
@@ -27,7 +27,8 @@ func classifyError(err error) string {
 	case errors.Is(err, chatsessions.ErrTargetEpisodeNotClosed),
 		errors.Is(err, chatsessions.ErrTargetEpisodeNumberExhausted),
 		errors.Is(err, chatsessions.ErrAttachmentBeyondStreamHead),
-		errors.Is(err, chatsessions.ErrAttachmentRetentionGap):
+		errors.Is(err, chatsessions.ErrAttachmentRetentionGap),
+		errors.Is(err, chatsessions.ErrUncommittedStreamPosition):
 		return "invariant_violation"
 	default:
 		return "validation"
