+++ pkg/services/chat_sessions/internal/service/session.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	chatsessions "github.com/portpowered/infinite-you/pkg/services/chat_sessions"
+	"github.com/portpowered/infinite-you/pkg/services/events"
 )
 
 // initialSessionVersion is the Version a newly created Session carries. A
@@ -67,13 +68,14 @@ func (s *Store) CreateSession(_ context.Context, req chatsessions.CreateSessionR
 	}
 
 	s.sessions[session.ID] = sessionRecord{
-		session:          session,
-		episodes:         []chatsessions.TargetEpisode{episode},
-		turns:            make(map[string]chatsessions.Turn),
-		turnsByRequest:   make(map[chatsessions.RequestIdentity]string),
-		attachments:      make(map[string]chatsessions.Attachment),
-		controls:         make(map[chatsessions.RequestIdentity]chatsessions.ControlIntent),
-		sequencedItemIDs: make(map[string]struct{}),
+		session:            session,
+		episodes:           []chatsessions.TargetEpisode{episode},
+		turns:              make(map[string]chatsessions.Turn),
+		turnsByRequest:     make(map[chatsessions.RequestIdentity]string),
+		attachments:        make(map[string]chatsessions.Attachment),
+		controls:           make(map[chatsessions.RequestIdentity]chatsessions.ControlIntent),
+		sequencedItemIDs:   make(map[string]struct{}),
+		sequencedPositions: make(map[events.AggregateSequence]sequencedSourceIdentity),
 	}
 	return chatsessions.CreateSessionResult{Session: session}, nil
 }
