+++ pkg/services/chat_sessions/internal/service/session_test.go
@@ -37,7 +37,7 @@ func validCreateRequest() chatsessions.CreateSessionRequest {
 func TestStore_CreateSession_Success(t *testing.T) {
 	ctx := context.Background()
 	now := time.Date(2026, 8, 2, 12, 0, 0, 0, time.UTC)
-	store := NewStore(sequentialIDs("session"), fixedClock(now))
+	store := NewStore(sequentialIDs("session"), fixedClock(now), nil, nil)
 
 	result, err := store.CreateSession(ctx, validCreateRequest())
 	if err != nil {
@@ -97,7 +97,7 @@ func TestStore_CreateSession_InvalidInputCreatesNoSession(t *testing.T) {
 	} {
 		t.Run(tt.name, func(t *testing.T) {
 			ctx := context.Background()
-			store := NewStore(sequentialIDs("session"), fixedClock(time.Now()))
+			store := NewStore(sequentialIDs("session"), fixedClock(time.Now()), nil, nil)
 
 			_, err := store.CreateSession(ctx, tt.mutate(validCreateRequest()))
 			if !errors.Is(err, tt.wantErr) {
@@ -117,7 +117,7 @@ func TestStore_CreateSession_InvalidInputCreatesNoSession(t *testing.T) {
 // Session cannot change what a later GetSession observes.
 func TestStore_CreateSession_ReturnsDetachedValue(t *testing.T) {
 	ctx := context.Background()
-	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()))
+	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()), nil, nil)
 
 	created, err := store.CreateSession(ctx, validCreateRequest())
 	if err != nil {
@@ -139,7 +139,7 @@ func TestStore_CreateSession_ReturnsDetachedValue(t *testing.T) {
 // collide on generated session identity.
 func TestStore_CreateSession_UniqueIDs(t *testing.T) {
 	ctx := context.Background()
-	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()))
+	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()), nil, nil)
 
 	first, err := store.CreateSession(ctx, validCreateRequest())
 	if err != nil {
@@ -156,7 +156,7 @@ func TestStore_CreateSession_UniqueIDs(t *testing.T) {
 
 func TestStore_GetSession_UnknownIDIsTypedNotFound(t *testing.T) {
 	ctx := context.Background()
-	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()))
+	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()), nil, nil)
 
 	_, err := store.GetSession(ctx, chatsessions.GetSessionRequest{SessionID: "does-not-exist"})
 	if !errors.Is(err, chatsessions.ErrNotFound) {
@@ -183,8 +183,8 @@ func TestStore_GetSession_UnknownIDIsTypedNotFound(t *testing.T) {
 // other.
 func TestStore_InstancesShareNoState(t *testing.T) {
 	ctx := context.Background()
-	first := NewStore(sequentialIDs("session"), fixedClock(time.Now()))
-	second := NewStore(sequentialIDs("session"), fixedClock(time.Now()))
+	first := NewStore(sequentialIDs("session"), fixedClock(time.Now()), nil, nil)
+	second := NewStore(sequentialIDs("session"), fixedClock(time.Now()), nil, nil)
 
 	created, err := first.CreateSession(ctx, validCreateRequest())
 	if err != nil {
@@ -206,7 +206,7 @@ func TestStore_InstancesShareNoState(t *testing.T) {
 // succeeds for every one with no cross-session interference.
 func TestStore_CreateSession_ConcurrentDifferentSessionsAreIndependent(t *testing.T) {
 	ctx := context.Background()
-	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()))
+	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()), nil, nil)
 
 	const n = 25
 	var wg sync.WaitGroup
