+++ pkg/services/chat_sessions/internal/service/store.go
@@ -56,24 +56,19 @@ type Store struct {
 	logger         logging.Logger
 }
 
-// WithEventsAppender sets the Events append port Sequence commits
-// source-native records through, and returns s for construction chaining.
-// It mutates s in place (Store embeds a sync.RWMutex, which must never be
-// copied) and is intended to be called once during construction wiring,
-// before s is shared across goroutines.
-func (s *Store) WithEventsAppender(appender EventsAppender) *Store {
-	s.eventsAppender = appender
-	return s
-}
-
-// WithEventsReader sets the Events read port AcknowledgeAttachment uses to
-// detect a retention gap between an attachment's current and requested
-// position, and returns s for construction chaining. It mutates s in place
-// (see WithEventsAppender) and is intended to be called once during
-// construction wiring, before s is shared across goroutines.
-func (s *Store) WithEventsReader(reader EventsReader) *Store {
-	s.eventsReader = reader
-	return s
+// sequencedSourceIdentity is the (SourceType, SourceID, SourceSequence,
+// SourceEventID) tuple Sequence committed for one aggregate position within
+// one session. AdvanceStreamHead consults sessionRecord.sequencedPositions
+// (keyed by that same aggregate position) against this exact tuple before
+// ever advancing StreamHead, so a caller can never move StreamHead to a
+// position this session's sequencer did not actually commit for the stated
+// source identity -- including a position that belongs to a different
+// session's own topic, or one committed for a different source tuple.
+type sequencedSourceIdentity struct {
+	SourceType     events.SourceType
+	SourceID       events.SourceID
+	SourceSequence events.SourceSequence
+	SourceEventID  events.SourceEventID
 }
 
 // sessionRecord is the Store-owned mutable aggregate for one Chat Session.
@@ -127,16 +122,24 @@ func (s *Store) WithEventsReader(reader EventsReader) *Store {
 // and adds to it only on a newly accepted record -- a duplicate resolution
 // never inserts a second time, since the ItemID it resolves to is already a
 // member from the original accepted call.
+// sequencedPositions holds, for every aggregate position this session's
+// sequencer has ever committed, the exact source identity tuple Sequence
+// committed it under. AdvanceStreamHead is the sole reader: it rejects any
+// requested AggregateSequence that is not a member here under the caller's
+// exact stated source identity, so StreamHead can never advance to a
+// fabricated, uncommitted, or cross-session position. Like
+// sequencedItemIDs, it is written only on a newly accepted Sequence record.
 type sessionRecord struct {
-	session          chatsessions.Session
-	episodes         []chatsessions.TargetEpisode
-	turns            map[string]chatsessions.Turn
-	turnsByRequest   map[chatsessions.RequestIdentity]string
-	lastTurnID       string
-	turnSequence     uint64
-	attachments      map[string]chatsessions.Attachment
-	controls         map[chatsessions.RequestIdentity]chatsessions.ControlIntent
-	sequencedItemIDs map[string]struct{}
+	session            chatsessions.Session
+	episodes           []chatsessions.TargetEpisode
+	turns              map[string]chatsessions.Turn
+	turnsByRequest     map[chatsessions.RequestIdentity]string
+	lastTurnID         string
+	turnSequence       uint64
+	attachments        map[string]chatsessions.Attachment
+	controls           map[chatsessions.RequestIdentity]chatsessions.ControlIntent
+	sequencedItemIDs   map[string]struct{}
+	sequencedPositions map[events.AggregateSequence]sequencedSourceIdentity
 }
 
 // activeTurnValue returns the session's current active Turn read live from
@@ -153,21 +156,29 @@ func (record sessionRecord) activeTurnValue() (chatsessions.Turn, bool) {
 }
 
 // NewStore constructs an empty Store from explicit dependencies. newID and
-// now must be non-nil. logger is optional and defaults to a no-op logger
-// when omitted, matching the repository's optional-logger construction
-// convention rather than a mutable reinjection path. Named NewStore (not
-// New) because this package also owns the unrelated FactoryTargetCatalog
+// now must be non-nil; eventsAppender and eventsReader may be nil for a
+// Store a caller knows will never exercise Sequence or AcknowledgeAttachment
+// (most existing Store tests predate the sequencer and never call either).
+// logger is optional and defaults to a no-op logger when omitted. Every
+// dependency is injected directly here, once, through this one canonical
+// constructor -- general-backend-standards.md §2 forbids a secondary
+// injector such as a post-construction `With<X>` mutator, so callers that do
+// need Sequence/AcknowledgeAttachment must supply both events ports up
+// front rather than attaching them afterward. Named NewStore (not New)
+// because this package also owns the unrelated FactoryTargetCatalog
 // Service's own New constructor (service.go); Go does not allow two
 // same-named top-level functions in one package.
-func NewStore(newID IDGenerator, now Clock, logger ...logging.Logger) *Store {
+func NewStore(newID IDGenerator, now Clock, eventsAppender EventsAppender, eventsReader EventsReader, logger ...logging.Logger) *Store {
 	var provided logging.Logger
 	if len(logger) > 0 {
 		provided = logger[0]
 	}
 	return &Store{
-		sessions: make(map[string]sessionRecord),
-		newID:    newID,
-		now:      now,
-		logger:   logging.EnsureLogger(provided),
+		sessions:       make(map[string]sessionRecord),
+		newID:          newID,
+		now:            now,
+		eventsAppender: eventsAppender,
+		eventsReader:   eventsReader,
+		logger:         logging.EnsureLogger(provided),
 	}
 }
