+++ pkg/services/chat_sessions/internal/service/target_test.go
@@ -23,7 +23,7 @@ func setTargetRequestID(id string) chatsessions.RequestIdentity {
 // for SetTarget calls.
 func newSetTargetTestSession(t *testing.T, now time.Time) (*Store, chatsessions.Session) {
 	t.Helper()
-	store := NewStore(sequentialIDs("session"), fixedClock(now))
+	store := NewStore(sequentialIDs("session"), fixedClock(now), nil, nil)
 	created, err := store.CreateSession(context.Background(), validCreateRequest())
 	if err != nil {
 		t.Fatalf("CreateSession: %v", err)
@@ -229,7 +229,7 @@ func TestStore_SetTarget_BusyWhileTurnActive(t *testing.T) {
 // an unknown SessionID reports *NotFoundError and mutates nothing.
 func TestStore_SetTarget_UnknownSessionIsTypedNotFound(t *testing.T) {
 	ctx := context.Background()
-	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()))
+	store := NewStore(sequentialIDs("session"), fixedClock(time.Now()), nil, nil)
 
 	_, err := store.SetTarget(ctx, chatsessions.SetTargetRequest{
 		RequestID:       setTargetRequestID("req-1"),
